estuary-cdk: re-organize `capture` module
This is pre-positioning for a future change that adds support for
refresh token rotation to `BaseCaptureConnector`. That change will
require moving `BaseCaptureConnector` out of `capture/__init__.py` to
resolve import paths, so I'm moving it now to make reviewing what's
actually changed easier. No functional changes have been made to
`BaseCaptureConnector` in this commit.

I've also re-organized the `capture` module to avoid breaking the
import path connectors are using to pull in `BaseCaptureConnector` from
the CDK. No functional changes in `task.py` either (it was previously
`__init__.py`), just some re-organization.
